mining: Move interfaces to mining/interface.go.
This moves the interfaces that are defined by mining to a separate file,
mining/interface.go.

This is part of an effort to split mining into several files to make the
package easier to navigate, and in preparation of adding additional test
coverage.
